+++ src/super_gradients/examples/regseg_transfer_learning_example/regseg_transfer_learning_example.py
@@ -1,8 +1,8 @@
 from super_gradients.training.datasets.dataset_interfaces.dataset_interface import SuperviselyPersonsDatasetInterface
 from super_gradients.training.sg_model import SgModel
 from super_gradients.training.metrics import BinaryIOU
-from super_gradients.training.utils.segmentation_utils import ColorJitterSeg, RandomFlip, RandomRescale, \
-    PadShortToCropSize, CropImageAndMask, ResizeSeg
+from super_gradients.training.transforms.transforms import ResizeSeg, RandomFlip, RandomRescale, CropImageAndMask, \
+    PadShortToCropSize, ColorJitterSeg
 from super_gradients.training.utils.callbacks import BinarySegmentationVisualizationCallback, Phase
 from torchvision import transforms
